add database flag and send properties
