+++ internal/flink/command_statement.go
@@ -4,8 +4,6 @@ import (
 	"time"
 
 	"github.com/spf13/cobra"
-
-	pcmd "github.com/confluentinc/cli/v3/pkg/cmd"
 )
 
 type statementOut struct {
@@ -30,18 +28,6 @@ func (c *command) newStatementCommand() *cobra.Command {
 	cmd.AddCommand(c.newStatementListCommand())
 
 	return cmd
-}
-
-func (c *command) addComputePoolFlag(cmd *cobra.Command) {
-	cmd.Flags().String("compute-pool", "", "Flink compute pool ID.")
-
-	pcmd.RegisterFlagCompletionFunc(cmd, "compute-pool", func(cmd *cobra.Command, args []string) []string {
-		if err := c.PersistentPreRunE(cmd, args); err != nil {
-			return nil
-		}
-
-		return c.autocompleteComputePools()
-	})
 }
 
 func (c *command) validStatementArgs(cmd *cobra.Command, args []string) []string {
