+++ internal/flink/command_statement_create.go
@@ -9,6 +9,7 @@ import (
 	pcmd "github.com/confluentinc/cli/v3/pkg/cmd"
 	"github.com/confluentinc/cli/v3/pkg/errors"
 	"github.com/confluentinc/cli/v3/pkg/flink"
+	"github.com/confluentinc/cli/v3/pkg/flink/config"
 	"github.com/confluentinc/cli/v3/pkg/output"
 )
 
@@ -23,6 +24,7 @@ func (c *command) newStatementCreateCommand() *cobra.Command {
 	cmd.Flags().String("sql", "", "The Flink SQL statement.")
 	c.addComputePoolFlag(cmd)
 	pcmd.AddServiceAccountFlag(cmd, c.AuthenticatedCLICommand)
+	c.addDatabaseFlag(cmd)
 	pcmd.AddEnvironmentFlag(cmd, c.AuthenticatedCLICommand)
 	pcmd.AddContextFlag(cmd, c.CLICommand)
 	pcmd.AddOutputFlag(cmd)
@@ -33,6 +35,16 @@ func (c *command) newStatementCreateCommand() *cobra.Command {
 }
 
 func (c *command) statementCreate(cmd *cobra.Command, args []string) error {
+	environmentId, err := c.Context.EnvironmentId()
+	if err != nil {
+		return err
+	}
+
+	environment, err := c.V2Client.GetOrgEnvironment(environmentId)
+	if err != nil {
+		return errors.NewErrorWithSuggestions(err.Error(), "List available environments with `confluent environment list`.")
+	}
+
 	computePool := c.Context.GetCurrentFlinkComputePool()
 	if computePool == "" {
 		return errors.NewErrorWithSuggestions("no compute pool selected", "Select a compute pool with `confluent flink compute-pool use` or `--compute-pool`.")
@@ -48,11 +60,22 @@ func (c *command) statementCreate(cmd *cobra.Command, args []string) error {
 		return err
 	}
 
+	database, err := cmd.Flags().GetString("database")
+	if err != nil {
+		return err
+	}
+
+	properties := map[string]string{config.ConfigKeyCatalog: environment.GetDisplayName()}
+	if database != "" {
+		properties[config.ConfigKeyDatabase] = database
+	}
+
 	statement := flinkgatewayv1beta1.SqlV1beta1Statement{
 		Name: flinkgatewayv1beta1.PtrString(name),
 		Spec: &flinkgatewayv1beta1.SqlV1beta1StatementSpec{
-			ComputePoolId: flinkgatewayv1beta1.PtrString(computePool),
 			Statement:     flinkgatewayv1beta1.PtrString(sql),
+			Properties:    &properties,
+			ComputePoolId: flinkgatewayv1beta1.PtrString(computePool),
 		},
 	}
 
@@ -65,11 +88,6 @@ func (c *command) statementCreate(cmd *cobra.Command, args []string) error {
 	if principal == "" {
 		output.ErrPrintln(flink.ServiceAccountWarning)
 		principal = c.Context.GetUser().GetResourceId()
-	}
-
-	environmentId, err := c.Context.EnvironmentId()
-	if err != nil {
-		return err
 	}
 
 	statement, err = client.CreateStatement(statement, principal, environmentId, c.Context.LastOrgId)
